requirements.txt
streamlit
pandas
plotly
requests
finnhub-python
openai
tiktoken
python-dotenv
